+++ pkgdown.yml
@@ -3,7 +3,7 @@ pkgdown: 2.2.0
 pkgdown_sha: ~
 articles:
   pvEBayes-introduction: pvEBayes-introduction.html
-last_built: 2026-03-17T18:49Z
+last_built: 2026-03-27T13:28Z
 urls:
   reference: https://yihaotancn.github.io/pvEBayes/reference
   article: https://yihaotancn.github.io/pvEBayes/articles
