### 2026-03-27 #760 @yabellini
Hi @YihaoTancn 👋

I’m the rOpenSci Community Manager, and I’d be happy to invite you to our Slack workspace if you’re interested. There’s absolutely no obligation to join—just an open invitation in case you’d like to connect with the community.
If you’d like an invite, feel free to email me at [[yabellini@ropensci.org](mailto:yabellini@ropensci.org)](mailto:[yabellini@ropensci.org](mailto:yabellini@ropensci.org)). And no rush at all—whenever it works for you!
