### 2026-04-07 #760 @collin-cademartori
---

### Compliance with Standards

- [x] This package complies with a sufficient number of standards for a silver badge
- [ ] This grade of badge is the same as what the authors wanted to achieve

The following standards currently deemed non-applicable (through tags of `@srrstatsNA`) could potentially be applied to future versions of this software: (Please specify)

I agree with the authors that all standards which were marked with `@srrstatsNA` were indeed not applicable to the package.

Please also comment on any standards which you consider either particularly well, or insufficiently, documented.

**General Standards:**

1. G2.14a error on missing data

    **Comment:** Based on my understanding of the implemented methods, erroring on missing data would be the only appropriate way to handle missingness. Some of this is handled by the internal function `.is_valid_contin_table` which gives a human-readable error when the input is not valid (including when NAs are present). However: (1) the error message could be more specific about why the input is not valid, and (2) at least one function (`estimate_null_expected_count`) performs separate validation testing outside of `.is_valid_contin_table` but does not provide a human-readable error in case of failure (instead using `stopifnot`). Probably `.is_valid_contin_table` should be used consistently throughout.

2. G2.16 All functions should also provide options to handle undefined values (e.g., `NaN`, `Inf` and `-Inf`), including potentially ignoring or removing such values.

    **Comment:** As before, it is my impression that the only reasonable way to handle undefined values is to error. The checks in `.is_valid_contin_table` error in case of `-Inf` and `NaN` entries, but not in case of `Inf` entries. It is not clear to me if this is intended behavior, but my guess is it is not, as it can lead to Inf values propagating to, e.g., expected counts.

3. G2.10 Software should ensure that extraction or filtering of single columns from tabular inputs should not presume any particular default behaviour, and should ensure all column-extraction operations behave consistently regardless of the class of tabular data used as input.

    **Comment:** The functions tagged as complying with this standard do not accept tabular inputs, and the authors' justification for compliance is not verbose enough to judge which lines within these two long functions the authors intended to support this.

4. G2.9 Software should issue diagnostic messages for type conversion in which information is lost (such as conversion of variables from factor to character; standardisation of variable names; or removal of meta-data such as those associated with [`sf`-format](https://r-spatial.github.io/sf/) data) or added (such as insertion of variable or column names where none were provided).

    **Comment:** Justification of compliance with this standard does not make sense to me. Also, if a user supplies a contingency table without row or column names to `pvEBayes` or `pvEBayes_tune`, default row and column names will be assigned without issuing any message to the user. This contingency table with new names will later be returned to the user in the output object, so the new names are potentially visible to the user.

5. G5.4 Correctness tests to test that statistical algorithms produce expected results to some fixed test data sets (potentially through comparisons using binding frameworks such as [RStata](https://github.com/lbraglia/RStata)).

    **Comment:** It appears that only the correctness of the `"general-gamma"` method is tested in the `test-correctness.R` file. Furthermore, the `"gps"` method is compared to an implementation in another package in an example provided in the documentation for the `pvEBayes` function, but this comparison is not carried out in the test suite as far as I can tell.

6. G5.7 Algorithm performance tests to test that implementation performs as expected as properties of data change. For instance, a test may show that parameters approach correct estimates within tolerance as data size increases, or that convergence times decrease for higher convergence thresholds.

    **Comment:** I could not find tests in the authors' test suite that matched this description. Test variations mostly involve introducing random noise or varying the input data, but none of these evaluated increasing/decreasing accuracy or speed of the algorithms as far as I could tell.

7. G5.8b Data of unsupported types (e.g., character or complex numbers in for functions designed only for numeric data)

    **Comment:** I did not see any tests for inputting data of the wrong type.

**Standards for Bayesian Software:**
1. BS1.4 For Bayesian Software which implements or otherwise enables convergence checkers, documentation should explicitly describe and provide examples of use with and without convergence checkers.

    **Comment:** This should arguably be tagged `@srrstatsNA`, since no MCMC samplers are used, and optimization tolerances are fundamentally different from convergence diagnostics for samplers. (E.g. optimization tolerances would not be turned off in any reasonable scenario.)

2. BS2.12 Bayesian Software should implement at least one parameter controlling the verbosity of output, defaulting to verbose output of all appropriate messages, warnings, errors, and progress indicators.

    **Comment:** Authors justify compliance by stating that only one function (for hyperparameter tuning) creates console output. However, this standard requires verbose output wherever appropriate, and arguably the optimization routines used for inference should produce output to indicate whether optimization succeeded. For instance, author's use third-party convex optimizer CXVR::psolve, which has a verbosity flag that the authors always set to `FALSE`. Arguably the posterior sampling functions should also at least offer progress printing.

3. BS2.2 Ensure that all appropriate validation and pre-processing of distributional parameters are implemented as distinct pre-processing steps prior to submitting to analytic routines, and especially prior to submitting to multiple parallel computational chains.

    **Comment:** Authors have tagged this with both `@srrstats` and `@srrstatsNA` in different places.

4. BS3.0 Explicitly document assumptions made in regard to missing values; for example that data is assumed to contain no missing (`NA`, `Inf`) values, and that such values, or entire rows including any such values, will be automatically removed from input data.

    **Comment:** Authors state in justification that no `NA` and `Inf` values occur in data to which `pvEBayes` applies in normal use, but that missing data is represented by, e.g., underreporting. This strikes me as tangential to the issue of handling `NA`s and `Inf`s, and the authors do not state where their claim is documented in the package. Arguably this could be a `@srrstatsNA` item.

5. BS5.3 Bayesian Software should return convergence statistics or equivalent

    **Comment:** Authors claim lack of use of MCMC means that software does not need to produce traditional convergence statistics. However, authors do rely on multiple optimization routines, at least several of which report (non)convergence of the optimizer. As far as I can tell, these convergence outputs are not checked, reported, or returned. Authors claim that optimizations error if convergence fails, though I was unable to test this.

6. BS5.5 Appropriate diagnostic statistics to indicate absence of convergence should either be returned or immediately able to be accessed.

    **Comment:** Same comment as in response to BS5.3

7. BS5.4 Where multiple checkers are enabled, Bayesian Software should return details of convergence checker used

    **Comment:** To my understanding, there is at most one convergence check performed per optimization routine. Therefore, it would make sense for this to be marked `@srrstatsNA`.

8. BS7.4a The implications of any assumptions on scales on input objects should be explicitly tested in this context; for example that the scales of inputs which do not have means of zero will not be able to be recovered.

    **Comment:** It is not clear to me which tests the authors intend to conform to this standard. My initial impression is that scaling of output with respect to input is not explicitly tested.

9. BS7.0 - BS7.2

    **Comment:** These three standards are tagged with both `@srrstats` and `@srrstatsNS` in different places.

For packages aiming for silver or gold badges:

- [x] This package extends beyond minimal compliance with standards in the following ways:

  1. **Demonstrating excellence in compliance with multiple standards from at least two broad sub-categories.** I agree with the authors that the package demonstrates excellent compliance with the general standards section 6.1.1. The package is very well documented. Documentation for both exported and internal functions is complete and detailed, and the provided vignette gives a useful introduction to the package's capabilities and intended use-cases. The authors also claim excellent compliance with 6.1.2, 6.1.5, and 6.2.5., and I believe the package could likely attain excellence in these subcategories if the above comments on particular standards are addressed.
  2. **Internal aspects of package structure and design.** I think the package reaches a level of notably excellent design in its implementation of 5 distinct empirical Bayes methods within a common interface for (1) input data and hyperparameters, (2) posterior sampling and visualization, and (3) hyperparameter tuning. The one area where I think improvement could be made is in reporting/exposing convergence indicators from underlying optimization algorithms (see comments on particular standards above).

---

### General Review

Overall, this package is thoughtfully designed, and it meaningfully advances the state of software tooling in its domain by providing high-quality implementations of competing inference methods unified by a common interface to facilitate easy comparisons. In its current state, I think the package meets the requirements for a silver badge, and a gold may be possible with further work. Two of the aspects along which a package may exceed the bronze level are addressed in the comments above. Regarding the other two aspects:

  1. **Compliance with a good number of standards beyond those identified as minimally necessary.** Reviewers did not specify which standards they thought were minimally necessary and potentially applicable, so I do not believe this aspect applies to this package.
  2. **Have a demonstrated generality of usage beyond one single envisioned use case.** For this aspect, I am not entirely sure what constitutes generality of use, and if the other reviewer and editor feel that this aspect is satisfied, then I am happy to defer to them. My current impression is that this aspect is not satisfied because (i) the package is focused on a specific inference strategy (empirical Bayes) applied to specific type of data (SRS tables), and (ii) the package appears to be mostly self-contained with limited interoperability with other packages, suggesting specificity rather than generality of usage.

- [x] **Community guidelines** including contribution guidelines in the README or CONTRIBUTING
- [x] The documentation is sufficient to enable general use of the package beyond one specific use case

#### Algorithms

- How well are algorithms encoded?

  Algorithms are well-implemented.

- Is the choice of computer language appropriate for that algorithm, and/or envisioned use of package?

  R is appropriate for high-level interface to inference algorithms, and lower-level languages are used for underlying optimization routines where appropriate.

- Are aspects of algorithmic scaling sufficiently documented and tested?

  I did not observe direct testing of algorithmic scaling. However, algorithms are demonstrated on various datasets of realistic size, which establishes that algorithm performance is sufficiently scalable for realistic use cases.

- Are there any aspects of algorithmic implementation which could be improved?

  - The package appears to implement a fixed null value for testing for a significant signal (1.001), as well as a fixed threshold for significance (0.95). It would be good to make these changeable by the user.

  - The `pvEBayes_tune` function only supports hyperparameter tuning for the methods "general-gamma" and "efron". If I understand the methods correctly, the "K-gamma" also has a hyperparameter (the integer K), but tuning is not supported for this hyperparameter. It might be useful if the authors could either implement tuning for this hyperparameter or explain why it is computationally infeasible or methodologically inappropriate to tune K.

  - The vignette presents a worked example using the `pvEBayes_tune` method with the "general-gamma" method. I attempted to reproduce the vignette with the "efron" method, but experienced two issues. First, with `method = "efron"`, `pvEBayes_tune` ran for about 5 minutes (compared to < 2 seconds with `method = "general-gamma"). Second, the method terminated with the error:
    ```
    Error in eval(call("force", as.symbol(paste0("..", x)))) :
    argument "..13" is missing, with no default
    ```
    I was not able to determine the cause of the error, but it would be good if the vignette example would also work with the "efron" method, ideally in a comparable amount of time if possible. Notably, using the "efron" method without tuning (with `pvEBayes` directly), I was able to get performance similar to that achieved with the "general-gamma" method.

  - After fitting a model with `fit <- pvEBayes(...)`, running `fit` in the R REPL prints nothing. This is in contradiction with the example in the `README` file, which suggests that running `fit` should print a concise model summary, i.e. as when one explicitly runs `print(fit)`.

  - The method `.KWDual_CVXR` iterates over three different solvers offered by `CVXR` and returns the first non-null solution. I did not see this behavior documented anywhere. It could be good to both (1) expose another function parameter to allow the user to explicitly choose which optimization routine they want to use and (2) document that the default behavior is to try the routines in the chosen order until a non-null result is found. Allowing the user to specify the optimization routine, in particular, will enable more controlled comparison of the proposed method with other methods.

- Regardless of actual coverage of tests, are there any fundamental software operations which are not sufficiently expressed in tests?

  As noted in the comment for standard G5.4, correctness tests appear to only test the correctness of one of the 5 implemented methods (the "General-Gamma" method). It would be useful to extend these tests to all implemented methods.

- Is there a need for extended tests, or if extended tests exists, have they been implemented in an appropriate way, and are they appropriately documented?

  There is no apparent need for extended tests.

#### Visualisation (where appropriate)

- Do visualisations aid the primary purposes of statistical interpretation of results?

  Visualizations generated by the `pvEBayes` class's `plot` method are detailed and well-designed for interpreting the output of the main inference algorithms. This is a strength of the package.

- Are there any aspects of visualisations which could risk statistical misinterpretation?

  There were no aspects of the visualizations which struck me as risking misinterpretation.

#### Package Design

- Is the package well designed for its intended purpose?

  The package does a very good job of implementing a well-documented, end-to-end pipeline for empirical Bayesian analysis of SRS tables. The custom summaries and visualizations are well-suited for interpreting the inference algorithm outputs.

- In relation to **External Design:** Do exported functions and the relationships between them enable general usage of the package?

  The external functions are mostly well-designed, with separate functions for different operations which can be composed to suit more general usage. It is useful, for instance, that the core optimization routines, posterior sampling, and hyperparameter tuning are separated into distinct functions. However, there is some inconsistency in what information can be obtained numerically versus visually. For example, the heatmap visualization prints posterior probabilities of a non-null signal between each drug-AE pair, but these probabilities do not appear to be returned by any of the summary functions, and are thus not easily obtained in numerical form.

- In relation to **External Design:** Do exported functions best serve inter-operability with other packages?

  For the most part, the package is self-contained, such that interfacing with other packages is unlikely to be necessary for regular use. Visualizations are provided as `ggplot` objects, making it easy for users to modify them as desired. I am unfamiliar with other packages dealing with SRS data, so I cannot speak to interoperability with any common data formats used in this field, if any.

- In relation to **Internal Design:** Are algorithms implemented appropriately in terms of aspects such as efficiency, flexibility, generality, and accuracy?

  Internal optimization and inference routines are complex numerical procedures, and I lack expertise in the particular empirical Bayes methods implemented. Therefore, I do not feel that I can confidently assess accuracy. At the same time, nothing in the implementations gave me reason to suspect any problems. The implemented routines are specific inference algorithms scoped to a particular class of models. In this sense, the algorithms are neither particularly general nor flexible, but this is appropriate given the intentions of the package. In particular, I see no reasonable means by which the algorithms could be made more general/flexible in a manner which immediately improves the usefulness of the package. As with accuracy, I am not well-equipped to assess efficiency except to note that the examples appear to run in a reasonable amount of time.

- In relation to **Internal Design:** Could ranges of admissible input structures, or form(s) of output structures, be expanded to enhance inter-operability with other packages?

  While admissible input structures are relatively constrained, this constraint seems to be natural and standard in the domain of SRS tables. I cannot imagine any way of expanding the admissible input structures in a manner which would meaningfully expand the usefulness of the package.

  One area where interoperability of output data structures could possibly be improved: Posterior samples are currently returned as an array in wide format, which makes it somewhat difficult to make custom plots using `ggplot` or otherwise process the samples with functions that expect data in long format. The conversion to long format appears to already take place within the internal plotting functions, so it would likely be easy to expose posterior samples in both wide and long format to the user.

#### Other Comments

  - In the `pvEBayes-introduction` vignette, I believe there is some inconsistency in notation between writing $\lambda$ (without subscripts) and $\lambda_{ij}$ (with subscripts) which makes it slightly unclear whether prior $g$ is over scalar components $\lambda_{ij}$ or jointly over the matrix $\lambda$. Seems to be the latter, but again, notation makes this somewhat ambiguous.
---
