### 2026-04-07 #760 @k-doering-NOAA
---

### Compliance with Standards

- [ ] This package complies with a sufficient number of standards for a gold badge
- [ ] This grade of badge is the same as what the authors wanted to achieve

In general, this package complies well with standards and has the potential to comply with all. In a few cases as listed below, the standard may need additional documentation to assure the package complies with all applicable standards (which I think is [required for Gold badges](https://stats-devguide.ropensci.org/pkgdev.html#pkgdev-gold)). With some minor changes, I suspect this package can comply with a sufficient number of(all applicable?)standards for a gold badge.

The following standards currently deemed non-applicable (through tags of `@srrstatsNA`) could potentially be applied to future versions of this software:

*BS4.3 Implement or otherwise offer at least one type of convergence checker, and provide a documented reference for that implementation.* - I’m not sure, but it seems like there must be a need to check for convergence still, even using an MLE algorithm to estimate the priors (empirical Bayes method)

*BS4.5 Ensure that appropriate mechanisms are provided for models which do not converge.* If it is true that a convergence checker needs to be provided for estimating the priors, then it seems reasonable that this standard should be met as well.

*BS7.3 Bayesian software should include tests which demonstrate and confirm the scaling of algorithmic efficiency with sizes of input data.* I think this standard could still apply to this package, even though the algorithms were developed and published in a separate study - or it should be thoroughly explained why it is not necessary (e.g., there is no expectation of difference in how the algorithms should perform based on the sizes of input data)

Please also comment on any standards which you consider either particularly well, or insufficiently, documented.

Many standards are well documented by the authors - they even documented multiple functions that met the standards. There were a few that could use more documentation, though:

*G2.3b Either: use tolower() or equivalent to ensure input of character parameters is not case dependent; or explicitly document that parameters are strictly case-sensitive.* `tolower()` is not used and the parameters are not explicitly documented as case sensitive.
*G5.3 For functions which are expected to return objects containing no missing (NA) or undefined (NaN, Inf) values, the absence of any such values in return objects should be explicitly tested.*  For this standard, I think the goal was to include checks before returning, rather than including a test in the test suite(although I could be wrong). I didn’t see any explicit checks for NA’s before returning from the pvEBayes main function. I also didn’t see an explicit check in the test that there were no missing or undefined values (if this standard actually meant to write a test to ensure that there were no missing or undefined values in return objects). In either interpretation of this standard, a change is needed.

G5.8 *Edge condition tests** *to test that these conditions produce expected behaviour such as clear warnings or errors when confronted with data with extreme properties including but not limited to:
*G5.8a Zero-length data*
*G5.8b Data of unsupported types (e.g., character or complex numbers in for functions designed only for numeric data)*
*G5.8c Data with all-NA fields or columns or all identical fields or columns*
*G5.8d Data outside the scope of the algorithm (for example, data with more fields (columns) than observations (rows) for some regression algorithms*

For all subbullets of 5.8 (G5.8a, G5.8b, G5.8c, G5.8d): I don’t think all of these types were tested in `test-pvEBayes_main_function.R`, but I think the 5.8 general standard of edge condition tests is met, but the specific subbullets are not. One suggestion is to add more signifiers in the script so it was clear what specifically was being tested for example, wrapping more statements in test_that() statements rather than every test in the  in 1 statement in the test-pvEBayes_main_function.R script; Include part of the specific error msg or warning msg expected to ensure the right message/error/warning is being caught in the test; or adding comments to make clear what error/warning/message is being tested. This could help for those performing maintenance on the package in the future, so that reasons for failing tests can be more quickly identified.

*BS1.0 Bayesian software which uses the term “hyperparameter” should explicitly clarify the meaning of that term in the context of that software.*
I did not see explanations of the term hyperparameter in the readme, vignette, or documentation.

*BS1.4 For Bayesian Software which implements or otherwise enables convergence checkers, documentation should explicitly describe and provide examples of use with and without convergence checkers.* If this standard does not apply to the package, it should be added to the srrstatsNA tag. Otherwise, documentation should be provided as specified in the standard.

*BS5.3 Bayesian Software should return convergence statistics or equivalent*
*BS5.4 Where multiple checkers are enabled, Bayesian Software should return details of convergence checker used*
*BS5.5 Appropriate diagnostic statistics to indicate absence of convergence should either be returned or immediately able to be accessed.*
I think BS5.3-5.5 should be moved to the srrstatsNA tag as the explanation is that they don’t apply to this software. Otherwise, if they do apply, the standards should be met.

G5.7 *Algorithm performance tests** to test that implementation performs as expected as properties of data change. For instance, a test may show that parameters approach correct estimates within tolerance as data size increases, or that convergence times decrease for higher convergence thresholds.*
I’m not sure I saw an algorithm performance test, but I’m also not sure if it applies to this package.

For packages aiming for silver or gold badges:

- [x] This package extends beyond minimal compliance with standards in the following ways: (please describe)

I agree that pvEBayes complies with a good number of standards overall, and that it has strong compliance with multiple standards across several broad categories. The pvEBayes package is also designed for use beyond a single envisioned application and demonstrates strong internal package design. I think once it has complied with all applicable standards, the pvEBayes package will be able to obtain a gold badge.

---

### General Review

Overall, this package is well formulated. Its scope is well defined, and it is thoroughly tested and documented. It complies with a high number of standards, although there are a few that could apply and a few that the explanation for meeting the standard was inadequate (details follow in the checklist below). I will note that reviewing for statistical soundness is not my expertise, and my familiarity with Empirical Bayes methods is minimal, so it is possible that some of my comments are not well formulated. The authors should not hesitate to correct me if I got anything wrong in the review!

Note there are a few checklist items that are not yet addressed by the package.

- [x] **Community guidelines** including contribution guidelines in the README or CONTRIBUTING
- [x] The documentation is sufficient to enable general use of the package beyond one specific use case
A statement of need is currently missing from the README.

#### Algorithms

I was not able to assess the statistical soundness of algorithms, as it is outside my area of expertise.

The readme mentions that open-source versions of algorithms (KM implementation) are used, which is a key feature of the package. It looks like the algorithm was copy/pasted from the pkg rather than the package used directly. Authors should check on included code to see if it follows the copyright guidance - https://devguide.ropensci.org/pkg_building.html#authorship-included-code

The package is generally well-tested. A few types of tests are missing per the standards section (see that section of the review for details).

#### Visualisation (where appropriate)

The visualizations provided by the “eyeplot” and “heatmap” types are both useful and appropriate for interpreting the results.

#### Package Design

The package is well designed for its intended purpose. There are 2 primary modeling functions for the user and it is easy to call them, as the inputs are consistent.

AIC and BIC functions are used to be consistent with other R packages. Summary and other standard s3 methods are used in this package to make it more familiar to users. The plotting functions work seamlessly with ggplot2. This package is interoperable with other R packages.

I did not assess if algorithms are implemented appropriately in terms of aspects such as efficiency, flexibility, generality, and accuracy, as this is outside my area of expertise.

#### Additional comments

A few small notes and suggestions for the authors’ consideration follow.

- Quick example: when I called fit, I did not see a precise description of the model, but instead I saw no output.
One idea is including the expected output on the README for the quick example, so the user is able to know if they have reproduced the quick example successfully or not.
- Vignette: “We consider an SRS dataset cataloging AE reports” - spell out AE at first mention.
In the vignette, there was one location where the formula notation did not render properly:  For illustration, we fit the model with a fixed hyperparameter value of $ = 0.5$ using the pvEBayes() function.
- I am confused why the [instructions in the vignette](https://yihaotancn.github.io/pvEBayes/articles/pvEBayes-introduction.html#fit-the-general-gamma-model) are different from the example in the documentation for the pvEBayes function. Why does the sim_table need to be created in the example from the documentation, but the statin2-25_44 data set can just be used directly for the contin_table input in the vignette?
- Small suggestion: For  model = "general-gamma", suggest listing all models in the usage (c(“general-gamma”, “efron”). Then, default to using the general-gamma using the match args function in the `pvEBayes()` and the `pvEBayes_tune()` functions.The current implementation seems fine since all possible model values are listed in the model arguments description. `Match.arg()` could also be considered in conjunction with this approach (as recommended by the r open sci packaging guide, which refers to this blog post: https://blog.r-hub.io/2022/03/10/input-checking/#checking-function-inputs-using-base-r
- Small suggestion for website: [grouping functions](https://devguide.ropensci.org/pkg_building.html#function-grouping) (as recommended by the ropensci packaging guide) could be helpful: https://yihaotancn.github.io/pvEBayes/reference/index.html

---
