+++ R/pvEBayes_main_function.R
@@ -34,11 +34,7 @@
 #'
 calculate_tilde_e <- function(contin_table) {
   contin_table <- as.matrix(contin_table)
-  stopifnot(
-    is.numeric(contin_table),
-    all(contin_table >= 0),
-    all(contin_table == floor(contin_table))
-  )
+  .is_valid_contin_table(contin_table)
   I <- nrow(contin_table)
   J <- ncol(contin_table)
   res <- contin_table[, J] %*% t(contin_table[I, ]) /
@@ -57,20 +53,94 @@ calculate_tilde_e <- function(contin_table) {
 #' @keywords internal
 #' @noRd
 .is_valid_contin_table <- function(contin_table) {
-  is_valid <- is.numeric(contin_table) &
-    all(contin_table >= 0) &
-    (all(contin_table == floor(contin_table)))
-
-  if (is_valid == FALSE) {
-    warning(
-      paste0(
-        "contin_table must be a matrix with",
-        " each entry being non-negative integer."
-      )
-    )
+  I <- nrow(contin_table)
+  J <- ncol(contin_table)
+  arg_name <- deparse(substitute(contin_table))
+  if (!is.numeric(contin_table)) {
+    stop("Invalid input: '%s' is not numeric." %>%
+      sprintf(arg_name))
   }
-  is_valid
+
+  if (any(is.na(contin_table))) {
+    stop("Invalid input: '%s' contains NA or NaN values." %>%
+      sprintf(arg_name))
+  }
+
+  if (any(is.infinite(contin_table))) {
+    stop("Invalid input: '%s' contains infinite (Inf/-Inf) values." %>%
+      sprintf(arg_name))
+  }
+
+  if (any(contin_table < 0)) {
+    stop("Invalid input: '%s' contains negative values." %>%
+      sprintf(arg_name))
+  }
+  if (any(rowSums(contin_table) == 0) | any(colSums(contin_table) == 0)) {
+    stop(paste0(
+      "Invalid input: '%s' contains at least",
+      " one row or column that is entirely zero."
+    ) %>%
+      sprintf(arg_name))
+  }
+  if (any(contin_table[I, ] == 0) | any(contin_table[, J] == 0)) {
+    stop(paste0(
+      "Invalid input: reference (I-th) row or (J-th) column in",
+      " '%s' contains at least one zero.",
+      " For detailed information, see the package vignette."
+    ) %>%
+      sprintf(arg_name))
+  }
+
+  if (any(contin_table != floor(contin_table))) {
+    stop("Invalid input: '%s' contains non-integer values." %>%
+      sprintf(arg_name))
+  }
+
+  TRUE
 }
+
+
+#' Check if a input is a valid positive numeric table
+#'
+#' @param x an IxJ numeric table
+#'
+#' @returns logical.
+#' @keywords internal
+#' @noRd
+.is_valid_numeric_table <- function(x) {
+  I <- nrow(x)
+  J <- ncol(x)
+  arg_name <- deparse(substitute(x))
+  if (!is.numeric(x)) {
+    stop("Invalid input: '%s' is not numeric." %>%
+      sprintf(arg_name))
+  }
+
+  if (any(is.na(x))) {
+    stop("Invalid input: '%s' contains NA or NaN values." %>%
+      sprintf(arg_name))
+  }
+
+  if (any(is.infinite(x))) {
+    stop("Invalid input: '%s' contains infinite (Inf/-Inf) values." %>%
+      sprintf(arg_name))
+  }
+
+  if (any(x < 0)) {
+    stop("Invalid input: '%s' contains negative values." %>%
+      sprintf(arg_name))
+  }
+  if (any(x == 0)) {
[…]
