+++ DESCRIPTION
@@ -43,7 +43,8 @@ Imports:
     wacolors,
     Rcpp,
     splines,
-    CVXR
+    CVXR,
+    cli
 LinkingTo: 
     Rcpp,
     RcppEigen

[…]
