+++ reference/extract_all_fitted_models.md
@@ -52,7 +52,8 @@ extract_all_fitted_models(tuned_object)
 #> General-gamma model with hyperparameter alpha = 0.1.
 #> Estimated prior is a mixture of 1 gamma distributions.
 #> 
-#> Running time of the general-gamma model fitting: 7e-04 seconds.
+#> Running time of the general-gamma model fitting: 8e-04 seconds.
+#> Optimizer convergence: successful.
 #> No posterior draws were generated.
 #> 
 #> Extract estimated prior parameters, discovered signals
@@ -64,7 +65,8 @@ extract_all_fitted_models(tuned_object)
 #> General-gamma model with hyperparameter alpha = 0.1.
 #> Estimated prior is a mixture of 1 gamma distributions.
 #> 
-#> Running time of the general-gamma model fitting: 7e-04 seconds.
+#> Running time of the general-gamma model fitting: 8e-04 seconds.
+#> Optimizer convergence: successful.
 #> No posterior draws were generated.
 #> 
 #> Extract estimated prior parameters, discovered signals
@@ -77,7 +79,8 @@ extract_all_fitted_models(tuned_object)
 #> General-gamma model with hyperparameter alpha = 0.
 #> Estimated prior is a mixture of 7 gamma distributions.
 #> 
-#> Running time of the general-gamma model fitting: 0.0064 seconds.
+#> Running time of the general-gamma model fitting: 0.0063 seconds.
+#> Optimizer convergence: not achieved.
 #> No posterior draws were generated.
 #> 
 #> Extract estimated prior parameters, discovered signals
@@ -89,7 +92,8 @@ extract_all_fitted_models(tuned_object)
 #> General-gamma model with hyperparameter alpha = 0.1.
 #> Estimated prior is a mixture of 1 gamma distributions.
 #> 
-#> Running time of the general-gamma model fitting: 7e-04 seconds.
+#> Running time of the general-gamma model fitting: 8e-04 seconds.
+#> Optimizer convergence: successful.
 #> No posterior draws were generated.
 #> 
 #> Extract estimated prior parameters, discovered signals
@@ -101,7 +105,8 @@ extract_all_fitted_models(tuned_object)
 #> General-gamma model with hyperparameter alpha = 0.3.
 #> Estimated prior is a mixture of 1 gamma distributions.
 #> 
-#> Running time of the general-gamma model fitting: 6e-04 seconds.
+#> Running time of the general-gamma model fitting: 7e-04 seconds.
+#> Optimizer convergence: successful.
 #> No posterior draws were generated.
 #> 
 #> Extract estimated prior parameters, discovered signals
@@ -113,7 +118,8 @@ extract_all_fitted_models(tuned_object)
 #> General-gamma model with hyperparameter alpha = 0.5.
 #> Estimated prior is a mixture of 1 gamma distributions.
 #> 
-#> Running time of the general-gamma model fitting: 7e-04 seconds.
+#> Running time of the general-gamma model fitting: 8e-04 seconds.
+#> Optimizer convergence: successful.
 #> No posterior draws were generated.
 #> 
 #> Extract estimated prior parameters, discovered signals
@@ -125,7 +131,8 @@ extract_all_fitted_models(tuned_object)
 #> General-gamma model with hyperparameter alpha = 0.7.
 #> Estimated prior is a mixture of 1 gamma distributions.
 #> 
-#> Running time of the general-gamma model fitting: 8e-04 seconds.
+#> Running time of the general-gamma model fitting: 9e-04 seconds.
+#> Optimizer convergence: successful.
 #> No posterior draws were generated.
 #> 
 #> Extract estimated prior parameters, discovered signals
[…]
