+++ articles/pvEBayes-introduction.md
@@ -144,6 +144,24 @@ gg_given_alpha <- pvEBayes(statin2025_44,
   model = "general-gamma",
   alpha = 0.5
 )
+#> ℹ Fitting general-gamma model...
+#> ✔ Fitting general-gamma model... [321ms]
+#> 
+#> ℹ Generating 1000 posterior draws...
+#> ✔ Generating 1000 posterior draws... [94ms]
+#> 
+#> Object of class 'pvEBayes'
+#> 
+#> General-gamma model with hyperparameter alpha = 0.5.
+#> Estimated prior is a mixture of 18 gamma distributions.
+#> 
+#> Running time of the general-gamma model fitting: 0.3302 seconds.
+#> Optimizer convergence: successful.
+#> Running time for posterior draws 
+#> (1000 signal strength posterior draws per AE-drug pair):0.196 seconds.
+#> 
+#> Extract estimated prior parameters, discovered signals
+#> and signal strength posterior draws using `summary()`.
 
 gg_given_alpha_detected_signal <- summary(gg_given_alpha,
   return = "detected signal"
