+++ reference/eyeplot_pvEBayes.md
@@ -86,6 +86,24 @@ fit <- pvEBayes(
   contin_table = statin2025_44, model = "general-gamma",
   alpha = 0.3, n_posterior_draws = 1000
 )
+#> ℹ Fitting general-gamma model...
+#> ✔ Fitting general-gamma model... [206ms]
+#> 
+#> ℹ Generating 1000 posterior draws...
+#> ✔ Generating 1000 posterior draws... [38ms]
+#> 
+#> Object of class 'pvEBayes'
+#> 
+#> General-gamma model with hyperparameter alpha = 0.3.
+#> Estimated prior is a mixture of 18 gamma distributions.
+#> 
+#> Running time of the general-gamma model fitting: 0.2136 seconds.
+#> Optimizer convergence: successful.
+#> Running time for posterior draws 
+#> (1000 signal strength posterior draws per AE-drug pair):0.046 seconds.
+#> 
+#> Extract estimated prior parameters, discovered signals
+#> and signal strength posterior draws using `summary()`.
 
 AE_names <- rownames(statin2025_44)[1:6]
 drug_names <- colnames(statin2025_44)[-7]
