+++ reference/print.pvEBayes.md
@@ -34,7 +34,36 @@ obj <- pvEBayes(
   contin_table = statin2025_44, model = "general-gamma",
   alpha = 0.5, n_posterior_draws = 10000
 )
+#> ℹ Fitting general-gamma model...
+#> ✔ Fitting general-gamma model... [302ms]
+#> 
+#> ℹ Generating 10000 posterior draws...
+#> ✔ Generating 10000 posterior draws... [477ms]
+#> 
+#> Object of class 'pvEBayes'
+#> 
+#> General-gamma model with hyperparameter alpha = 0.5.
+#> Estimated prior is a mixture of 18 gamma distributions.
+#> 
+#> Running time of the general-gamma model fitting: 0.3098 seconds.
+#> Optimizer convergence: successful.
+#> Running time for posterior draws 
+#> (10000 signal strength posterior draws per AE-drug pair):0.4849 seconds.
+#> 
+#> Extract estimated prior parameters, discovered signals
+#> and signal strength posterior draws using `summary()`.
 
 print(obj)
+#> Object of class 'pvEBayes'
 #> 
+#> General-gamma model with hyperparameter alpha = 0.5.
+#> Estimated prior is a mixture of 18 gamma distributions.
+#> 
+#> Running time of the general-gamma model fitting: 0.3098 seconds.
+#> Optimizer convergence: successful.
+#> Running time for posterior draws 
+#> (10000 signal strength posterior draws per AE-drug pair):0.4849 seconds.
+#> 
+#> Extract estimated prior parameters, discovered signals
+#> and signal strength posterior draws using `summary()`.
 ```

[…]
