+++ reference/posterior_draws.md
@@ -43,14 +43,14 @@ fit <- pvEBayes(
   alpha = 0.3, n_posterior_draws = NULL
 )
 #> ℹ Fitting general-gamma model...
-#> ✔ Fitting general-gamma model... [208ms]
+#> ✔ Fitting general-gamma model... [225ms]
 #> 
 #> Object of class 'pvEBayes'
 #> 
 #> General-gamma model with hyperparameter alpha = 0.3.
 #> Estimated prior is a mixture of 18 gamma distributions.
 #> 
-#> Running time of the general-gamma model fitting: 0.216 seconds.
+#> Running time of the general-gamma model fitting: 0.2322 seconds.
 #> Optimizer convergence: successful.
 #> No posterior draws were generated.
 #> 
