+++ reference/extract_all_fitted_models.md
@@ -79,7 +79,7 @@ extract_all_fitted_models(tuned_object)
 #> General-gamma model with hyperparameter alpha = 0.
 #> Estimated prior is a mixture of 7 gamma distributions.
 #> 
-#> Running time of the general-gamma model fitting: 0.0063 seconds.
+#> Running time of the general-gamma model fitting: 0.0062 seconds.
 #> Optimizer convergence: not achieved.
 #> No posterior draws were generated.
 #> 
@@ -131,7 +131,7 @@ extract_all_fitted_models(tuned_object)
 #> General-gamma model with hyperparameter alpha = 0.7.
 #> Estimated prior is a mixture of 1 gamma distributions.
 #> 
-#> Running time of the general-gamma model fitting: 9e-04 seconds.
+#> Running time of the general-gamma model fitting: 8e-04 seconds.
 #> Optimizer convergence: successful.
 #> No posterior draws were generated.
 #> 

[…]
