+++ R/pvEBayes_main_function.R
@@ -591,8 +591,8 @@ estimate_null_expected_count <- function(contin_table) {
 #' @keywords internal
 .E_fit <- function(N, E, c0 = 1, pDegree = 5, aStart = 1, rel.tol) {
   n_draws <- prod(dim(N)) * 1.5
-  if (n_draws >= 400) {
-    n_draws <- 400
+  if (n_draws >= 500) {
+    n_draws <- 500
   }
   if (n_draws < 50) {
     n_draws <- 50
@@ -1029,13 +1029,15 @@ estimate_null_expected_count <- function(contin_table) {
 #'
 #' Methods "K-gamma" and "general-gamma" are non-parametric empirical Bayes
 #' models, introduced by Tan et al. (2025). The number of mixture components "K"
-#' needs to be prespecified when fitting a "K-gamma" model. For "general-gamma",
-#' the mixture weights are regularized by a Dirichlet hyper prior with
-#' hyperparameter \eqn{0 \leq \alpha < 1} that controls the shrinkage strength.
-#' As "alpha" approaches 0, less non-empty mixture components exist in the
-#' fitted model. When \eqn{\alpha = 0}, the Dirichlet distribution is an
-#' improper prior still offering a reasonable posterior inference that
-#' represents the strongest shrinkage of the "general-gamma" model.
+#' needs to be prespecified when fitting a "K-gamma" model. When the number of
+#' mixture components is unknown, we recommend using the "general-gamma" model
+#' instead. For "general-gamma", the mixture weights are regularized by a
+#' Dirichlet hyper prior with hyperparameter \eqn{0 \leq \alpha < 1} that
+#' controls the shrinkage strength. As "alpha" approaches 0, less non-empty
+#' mixture components exist in the fitted model. When \eqn{\alpha = 0}, the
+#' Dirichlet distribution is an improper prior still offering a reasonable
+#' posterior inference that represents the strongest shrinkage of the
+#' "general-gamma" model.
 #'
 #' Parameter estimation for the "KM" model is formulated as a convex
 #' optimization problem. The objective function and constraints follow the same
@@ -1437,7 +1439,7 @@ pvEBayes <- function(contin_table, model = "general-gamma",
 #' will be used.
 #' @param p_vec vector of hyperparameter p values to be selected. p is
 #' a hyperparameter in "efron" model which should be a positive integer. If is
-#' NULL, a default set of p values (80, 100, 120, 150, 200) will be used.
+#' NULL, a default set of p values (40, 60, 80, 100, 120) will be used.
 #' @param c0_vec vector of hyperparameter c0 values to be selected. c0 is
 #' a hyperparameter in "efron" model which should be a positive number. If is
 #' NULL, a default set of c0 values (0.001, 0.01, 0.1, 0.2, 0.5) will be used.
@@ -1772,7 +1774,7 @@ tuning_general_gamma <- function(contin_table,
 #' adverse events for I AEs (along the rows) and J drugs (along the columns).
 #' @param p_vec vector of hyperparameter p values to be selected. p is
 #' a hyperparameter in "efron" model which should be a positive integer. If is
-#' NULL, a default set of p values (80, 100, 120, 150, 200) will be used.
+#' NULL, a default set of p values (40, 60, 80, 100, 120) will be used.
 #' @param c0_vec vector of hyperparameter c0 values to be selected. c0 is
 #' a hyperparameter in "efron" model which should be a positive number. If is
 #' NULL, a default set of c0 values (0.001, 0.01, 0.1, 0.2, 0.5) will be used.
