+++ man/get_posterior_prob.Rd
@@ -1,0 +1,22 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/pvEBayes_object_S3_methods.R
+\name{get_posterior_prob}
+\alias{get_posterior_prob}
+\title{Obtain posterior probability of being a signal}
+\usage{
+get_posterior_prob(obj, cutoff_signal = 1.001)
+}
+\arguments{
+\item{obj}{a \code{pvEBayes} object, which is the output of the function
+\link{pvEBayes} or \link{pvEBayes_tune}.}
+
+\item{cutoff_signal}{numeric. Threshold for signal detection. An AE-drug
+combination is classified as a detected signal if its 5th posterior
+percentile exceeds this threshold.}
+}
+\value{
+a matrix
+}
+\description{
+Obtain posterior probability of being a signal
+}
