+++ NAMESPACE
@@ -11,6 +11,7 @@ export(estimate_null_expected_count)
 export(extract_all_fitted_models)
 export(eyeplot_pvEBayes)
 export(generate_contin_table)
+export(get_posterior_prob)
 export(heatmap_pvEBayes)
 export(posterior_draws)
 export(pvEBayes)

[…]
