+++ reference/eyeplot_pvEBayes.md
@@ -97,7 +97,7 @@ fit <- pvEBayes(
 #> General-gamma model with hyperparameter alpha = 0.3.
 #> Estimated prior is a mixture of 18 gamma distributions.
 #> 
-#> Running time of the general-gamma model fitting: 0.2285 seconds.
+#> Running time of the general-gamma model fitting: 0.229 seconds.
 #> Optimizer convergence: successful.
 #> Running time for posterior draws 
 #> (1000 signal strength posterior draws per AE-drug pair):0.0461 seconds.
