+++ reference/AIC.pvEBayes.md
@@ -48,7 +48,7 @@ fit <- pvEBayes(
 #> General-gamma model with hyperparameter alpha = 0.3.
 #> Estimated prior is a mixture of 18 gamma distributions.
 #> 
-#> Running time of the general-gamma model fitting: 0.2395 seconds.
+#> Running time of the general-gamma model fitting: 0.2387 seconds.
 #> Optimizer convergence: successful.
 #> No posterior draws were generated.
 #> 
