+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: pvEBayes
 Title: Empirical Bayes Methods for Pharmacovigilance
-Version: 0.2.2
+Version: 0.3.0
 Authors@R: c(
     person(
     "Yihao", "Tan", 
