+++ reference/plot.pvEBayes.md
@@ -39,22 +39,23 @@ A [ggplot](https://ggplot2.tidyverse.org/reference/ggplot.html) object.
 ## Examples
 
 ``` r
+
 obj <- pvEBayes(statin2025_44, model = "general-gamma", alpha = 0.5)
 #> ℹ Fitting general-gamma model...
-#> ✔ Fitting general-gamma model... [323ms]
+#> ✔ Fitting general-gamma model... [314ms]
 #> 
 #> ℹ Generating 1000 posterior draws...
-#> ✔ Generating 1000 posterior draws... [37ms]
+#> ✔ Generating 1000 posterior draws... [35ms]
 #> 
 #> Object of class 'pvEBayes'
 #> 
 #> General-gamma model with hyperparameter alpha = 0.5.
 #> Estimated prior is a mixture of 18 gamma distributions.
 #> 
-#> Running time of the general-gamma model fitting: 0.3307 seconds.
+#> Running time of the general-gamma model fitting: 0.3222 seconds.
 #> Optimizer convergence: successful.
 #> Running time for posterior draws 
-#> (1000 signal strength posterior draws per AE-drug pair):0.0439 seconds.
+#> (1000 signal strength posterior draws per AE-drug pair):0.0422 seconds.
 #> 
 #> Extract estimated prior parameters, discovered signals
 #> and signal strength posterior draws using `summary()`.
