+++ reference/posterior_draws.md
@@ -38,19 +38,20 @@ draws.
 ## Examples
 
 ``` r
+
 fit <- pvEBayes(
   contin_table = statin2025_44, model = "general-gamma",
   alpha = 0.3, n_posterior_draws = NULL
 )
 #> ℹ Fitting general-gamma model...
-#> ✔ Fitting general-gamma model... [217ms]
+#> ✔ Fitting general-gamma model... [212ms]
 #> 
 #> Object of class 'pvEBayes'
 #> 
 #> General-gamma model with hyperparameter alpha = 0.3.
 #> Estimated prior is a mixture of 18 gamma distributions.
 #> 
-#> Running time of the general-gamma model fitting: 0.224 seconds.
+#> Running time of the general-gamma model fitting: 0.2197 seconds.
 #> Optimizer convergence: successful.
 #> No posterior draws were generated.
 #> 
@@ -59,6 +60,6 @@ fit <- pvEBayes(
 
 fit_with_draws <- posterior_draws(fit, n_posterior_draws = 1000)
 #> ℹ Generating 1000 posterior draws...
-#> ✔ Generating 1000 posterior draws... [36ms]
+#> ✔ Generating 1000 posterior draws... [35ms]
 #> 
 ```
