+++ articles/pvEBayes-introduction.md
@@ -7,56 +7,69 @@ collection of spontaneous reports of suspected adverse drug events from
 pharmaceutical companies, healthcare professionals, and patients. These
 reports are curated and stored in spontaneous reporting systems (SRS),
 usually organized as a large frequency table for downstream data
-analysis. We consider an SRS dataset cataloging AE reports on $I$ AE
-rows across $J$ drug columns. Let $N_{ij}$ denote the number of reported
-cases for the $i$-th AE and the $j$-th drug, where $i = 1,...,I$ and
-$j = 1,...,J$. Therefore, AE-drug pairwise occurrences from the
-AE-reports are summarized into an $I \times J$ contingency table, where
-the $(i,j)$-th cell catalogs the observed count $N_{ij}$ indicating the
-number of cases involving $i$-th AE and the $j$-th drug.
+analysis. We consider an SRS dataset cataloging AE reports on $`I`$ AE
+rows across $`J`$ drug columns. Let $`N_{ij}`$ denote the number of
+reported cases for the $`i`$-th AE and the $`j`$-th drug, where
+$`i = 1,..., I`$ and $`j = 1,..., J`$. Therefore, AE-drug pairwise
+occurrences from the AE-reports are summarized into an $`I \times J`$
+contingency table, where the $`(i,j)`$-th cell catalogs the observed
+count $`N_{ij}`$ indicating the number of cases involving $`i`$-th AE
+and the $`j`$-th drug.
 
 Current disproportionality analysis mainly focuses on *signal detection*
-which seeks to determine whether the observation $N_{ij}$ is
-substantially greater than the corresponding null baseline $E_{ij}$. Or
-equivalently, its signal strength $\lambda_{ij}$ is significantly
+which seeks to determine whether the observation $`N_{ij}`$ is
+substantially greater than the corresponding null baseline $`E_{ij}`$.
+Or equivalently, its signal strength $`\lambda_{ij}`$ is significantly
 greater than 1.
 
 In addition to *signal detection*, Tan et al. (*Stat. in Med.*, 2025)
 broaden the role of disproportionality to *signal strength estimation*.
 The use of the flexible non-parametric empirical Bayes models enables
 more nuanced empirical Bayes posterior inference (signal strength
-$\lambda$ estimation and uncertainty quantification). This allows
-researchers to distinguish AE-drug pairs that would appear similar under
-a binary signal detection framework. For example, the AE-drug pairs with
-signal strengths of 1.5 and 4.0 could both be significantly greater than
-1 and detected as a signal. Such differences in signal strength may have
-distinct implications in medical and clinical contexts.
+$`\{\lambda_{ij}\}`$ estimation and uncertainty quantification). This
+allows researchers to distinguish AE-drug pairs that would appear
+similar under a binary signal detection framework. For example, the
+AE-drug pairs with signal strengths of 1.5 and 4.0 could both be
+significantly greater than 1 and detected as a signal. Such differences
+in signal strength may have distinct implications in medical and
+clinical contexts.
 
-Methods implemented in this package assumes the observed count $N_{ij}$
-conditional on $E_{ij}$ is that
-$$N_{ij} \mid E_{ij} \sim \operatorname{Poisson}\left( E_{ij}\lambda_{ij} \right),\ N_{ij} = 0,1,2\ldots,$$
[…]
