+++ reference/BIC.pvEBayes.md
@@ -31,19 +31,20 @@ numeric, BIC score for the resulting model.
 ## Examples
 
 ``` r
+
 fit <- pvEBayes(
   contin_table = statin2025_44, model = "general-gamma",
   alpha = 0.3, n_posterior_draws = NULL
 )
 #> ℹ Fitting general-gamma model...
-#> ✔ Fitting general-gamma model... [228ms]
+#> ✔ Fitting general-gamma model... [212ms]
 #> 
 #> Object of class 'pvEBayes'
 #> 
 #> General-gamma model with hyperparameter alpha = 0.3.
 #> Estimated prior is a mixture of 18 gamma distributions.
 #> 
-#> Running time of the general-gamma model fitting: 0.2448 seconds.
+#> Running time of the general-gamma model fitting: 0.2285 seconds.
 #> Optimizer convergence: successful.
 #> No posterior draws were generated.
 #> 
