+++ reference/eyeplot_pvEBayes.md
@@ -87,7 +87,7 @@ fit <- pvEBayes(
   alpha = 0.3, n_posterior_draws = 1000
 )
 #> ℹ Fitting general-gamma model...
-#> ✔ Fitting general-gamma model... [215ms]
+#> ✔ Fitting general-gamma model... [216ms]
 #> 
 #> ℹ Generating 1000 posterior draws...
 #> ✔ Generating 1000 posterior draws... [37ms]
@@ -97,10 +97,10 @@ fit <- pvEBayes(
 #> General-gamma model with hyperparameter alpha = 0.3.
 #> Estimated prior is a mixture of 18 gamma distributions.
 #> 
-#> Running time of the general-gamma model fitting: 0.2231 seconds.
+#> Running time of the general-gamma model fitting: 0.2238 seconds.
 #> Optimizer convergence: successful.
 #> Running time for posterior draws 
-#> (1000 signal strength posterior draws per AE-drug pair):0.0446 seconds.
+#> (1000 signal strength posterior draws per AE-drug pair):0.0445 seconds.
 #> 
 #> Extract estimated prior parameters, discovered signals
 #> and signal strength posterior draws using `summary()`.
