+++ articles/pvEBayes-introduction.md
@@ -152,33 +152,138 @@ estimating the corresponding signal strength of these 6 statin drugs.
 These are achieved by fitting empirical Bayes models with pvEBayes() to
 the SRS data. We begin by fitting the general-gamma model to this
 dataset. Other models mentioned above could be used by modifying the
-‘model’ argument. For illustration, we fit the model with a fixed
-hyperparameter value of \$ = 0.5\$ using the pvEBayes() function.
+‘model’ argument. “general-gamma” method has a hyperparameter
+$`\alpha \in [0,1]`$. Small $`\alpha`$ encourages shrinkage on mixture
+weights of the estimated prior distribution. See Tan et al. (2025, ) for
+further details. For illustration, we fit the model with a fixed
+hyperparameter value of $`\alpha = 0.5`$ using the pvEBayes() function.
 
 ``` r
 
 gg_given_alpha <- pvEBayes(statin2025_44,
   model = "general-gamma",
-  alpha = 0.5
+  alpha = 0.5,
+  maxi = 200,
+  tol_ecm = 1e-4  #default value
 )
 #> ℹ Fitting general-gamma model...
-#> ✔ Fitting general-gamma model... [334ms]
+#> ✔ Fitting general-gamma model... [335ms]
 #> 
 #> ℹ Generating 1000 posterior draws...
-#> ✔ Generating 1000 posterior draws... [96ms]
+#> ✔ Generating 1000 posterior draws... [92ms]
 #> 
 #> Object of class 'pvEBayes'
 #> 
 #> General-gamma model with hyperparameter alpha = 0.5.
 #> Estimated prior is a mixture of 18 gamma distributions.
 #> 
-#> Running time of the general-gamma model fitting: 0.3433 seconds.
+#> Running time of the general-gamma model fitting: 0.3439 seconds.
 #> Optimizer convergence: successful.
 #> Running time for posterior draws 
-#> (1000 signal strength posterior draws per AE-drug pair):0.1901 seconds.
+#> (1000 signal strength posterior draws per AE-drug pair):0.1866 seconds.
 #> 
 #> Extract estimated prior parameters, discovered signals
 #> and signal strength posterior draws using `summary()`.
+
+gg_given_alpha$iter
+#> [1] 167
+```
+
+The convergence criterion in
+[`pvEBayes()`](https://yihaotancn.github.io/pvEBayes/reference/pvEBayes.md)
+is based on the relative change in the log likelihood between
+consecutive iterations. In this example, the algorithm stopped after
+satisfying the convergence criterion. This is also reflected in the
+printed model fitting message, where the optimizer convergence status is
+reported as successful. We can inspect the number of iterations used by
+the fitting algorithm by `gg_given_alpha\$iter`. In this fit,
+convergence was achieved after 167 iterations.
+
+The next example illustrates a case in which the algorithm stops because
+it reaches the maximum number of iterations. Here, we use a smaller
+convergence tolerance.
+
+``` r
+
+
+
+gg_given_alpha2 <- pvEBayes(statin2025_44,
+  model = "general-gamma",
+  alpha = 0.5,
+  maxi = 200,
+  tol_ecm = 1e-8  #smaller tolerance for convergence 
+)
+#> ℹ Fitting general-gamma model...
+#> ✔ Fitting general-gamma model... [346ms]
+#> 
+#> ℹ Generating 1000 posterior draws...
+#> ✔ Generating 1000 posterior draws... [35ms]
+#> 
+#> Object of class 'pvEBayes'
+#> 
+#> General-gamma model with hyperparameter alpha = 0.5.
+#> Estimated prior is a mixture of 18 gamma distributions.
+#> 
+#> Running time of the general-gamma model fitting: 0.3544 seconds.
[…]
