### 2026-05-03 #760 @maurolepore
I see both reviews are in and @LucyMcGowan is steering this through. Looking forward to your response to each review.

Thanks for sharing your work with the rOpenSci community.
