+++ R/pvEBayes_main_function.R
@@ -398,6 +398,15 @@ estimate_null_expected_count <- function(contin_table) {
 #' @param ... additional parameters to be passed to optimizer for 'KM' model.
 #' See 'CVXR::solve' for detail.
 #'
+#' @details
+#'
+#' Parameter estimation for the "KM" model is formulated as a convex
+#' optimization problem. The objective function and constraints are modified
+#' from the \pkg{REBayes} package (see 'REBayes::KWDual()'). Parameter
+#' estimation is performed using the open-source convex optimization package
+#' \pkg{CVXR}.
+#'
+#'
 #' @returns a list of CVXR optimizer outputs
 #' @keywords internal
 #' @noRd
@@ -1042,8 +1051,8 @@ estimate_null_expected_count <- function(contin_table) {
 #' "general-gamma" model.
 #'
 #' Parameter estimation for the "KM" model is formulated as a convex
-#' optimization problem. The objective function and constraints follow the same
-#' construction as in \pkg{REBayes} (see 'REBayes::KWDual()'). Parameter
+#' optimization problem. The objective function and constraints are modified
+#' from the \pkg{REBayes} package (see 'REBayes::KWDual()'). Parameter
 #' estimation is performed using the open-source convex optimization package
 #' \pkg{CVXR}.
 #'
@@ -1103,33 +1112,18 @@ estimate_null_expected_count <- function(contin_table) {
 #' @examples
 #'
 #' set.seed(1)
-#' ref_table <- statin2025_44
-#'
-#' # set up signal matrix with one signal at entry (1,1)
-#' sig_mat <- matrix(1, nrow(ref_table), ncol(ref_table))
-#' sig_mat[c(1, 5), 1] <- 2
-#'
-#'
-#' simu_table <- generate_contin_table(
-#'   ref_table = ref_table,
-#'   signal_mat = sig_mat,
-#'   n_table = 1
-#' )[[1]]
-#'
 #'
 #' # fit general-gamma model with a specified alpha
 #' fit <- pvEBayes(
-#'   contin_table = simu_table,
+#'   contin_table = statin2025_44,
 #'   model = "general-gamma",
 #'   alpha = 0.3,
-#'   n_posterior_draws = 1000,
-#'   E = NULL,
-#'   maxi = NULL
+#'   n_posterior_draws = 1000
 #' )
 #'
 #' # fit K-gamma model with K = 3
 #' fit_Kgamma <- pvEBayes(
-#'   contin_table = simu_table, model = "K-gamma",
+#'   contin_table = statin2025_44, model = "K-gamma",
 #'   K = 3, n_posterior_draws = 1000
 #' )
 #'
@@ -1137,20 +1131,18 @@ estimate_null_expected_count <- function(contin_table) {
 #' # fit Efron model with specified hyperparameters
 #' # p = 40, c0 = 0.05
 #'
-#' \dontrun{
 #' fit_efron <- pvEBayes(
-#'   contin_table = simu_table,
+#'   contin_table = statin2025_44,
 #'   model = "efron",
 #'   p = 40,
 #'   c0 = 0.05,
 #'   n_posterior_draws = 1000
 #' )
-#' }
 #'
 #' # fit GPS model and comapre with 'openEBGM'
 #'
 #'
-#' fit_gps <- pvEBayes(simu_table, model = "GPS")
+#' fit_gps <- pvEBayes(statin2025_44, model = "GPS")
 #'
 #' \dontrun{
 #'
@@ -1159,16 +1151,16 @@ estimate_null_expected_count <- function(contin_table) {
 #' ## tol_ecm is the absolute tolerance for ECM stopping rule.
 #' ## It is set to ensure comparability to `openEBGM`.
 #'
[…]
