+++ reference/pvEBayes.md
@@ -10,7 +10,7 @@ specified hyperparameter, if is required. Supported models include the
 ``` r
 pvEBayes(
   contin_table,
-  model = "general-gamma",
+  model = c("general-gamma", "K-gamma", "GPS", "KM", "efron"),
   alpha = NULL,
   K = NULL,
   p = NULL,
@@ -51,7 +51,8 @@ pvEBayes(
 
   a integer greater than or equal to 2 indicating the number of mixture
   components in the prior distribution. It is needed if "K-gamma" model
-  is used. See Tan et al. (2025) for further details.
+  is used. When K is unknown, please consider its extension
+  "general-gamma" instead. See Tan et al. (2025) for further details.
 
 - p:
 
@@ -148,8 +149,8 @@ prior still offering a reasonable posterior inference that represents
 the strongest shrinkage of the "general-gamma" model.
 
 Parameter estimation for the "KM" model is formulated as a convex
-optimization problem. The objective function and constraints follow the
-same construction as in REBayes (see 'REBayes::KWDual()'). Parameter
+optimization problem. The objective function and constraints are
+modified from the REBayes package (see 'REBayes::KWDual()'). Parameter
 estimation is performed using the open-source convex optimization
 package CVXR.
 
@@ -193,67 +194,52 @@ Optimization. *Journal of Statistical Software*. 2020; 94;14:1-34.
 ``` r
 
 set.seed(1)
-ref_table <- statin2025_44
-
-# set up signal matrix with one signal at entry (1,1)
-sig_mat <- matrix(1, nrow(ref_table), ncol(ref_table))
-sig_mat[c(1, 5), 1] <- 2
-
-
-simu_table <- generate_contin_table(
-  ref_table = ref_table,
-  signal_mat = sig_mat,
-  n_table = 1
-)[[1]]
-
 
 # fit general-gamma model with a specified alpha
 fit <- pvEBayes(
-  contin_table = simu_table,
+  contin_table = statin2025_44,
   model = "general-gamma",
   alpha = 0.3,
-  n_posterior_draws = 1000,
-  E = NULL,
-  maxi = NULL
+  n_posterior_draws = 1000
 )
 #> ℹ Fitting general-gamma model...
-#> ✔ Fitting general-gamma model... [196ms]
+#> ✔ Fitting general-gamma model... [211ms]
 #> 
 #> ℹ Generating 1000 posterior draws...
-#> ✔ Generating 1000 posterior draws... [36ms]
+#> ✔ Generating 1000 posterior draws... [35ms]
 #> 
 #> Object of class 'pvEBayes'
 #> 
 #> General-gamma model with hyperparameter alpha = 0.3.
-#> Estimated prior is a mixture of 4 gamma distributions.
+#> Estimated prior is a mixture of 18 gamma distributions.
 #> 
-#> Running time of the general-gamma model fitting: 0.2043 seconds.
+#> Running time of the general-gamma model fitting: 0.2192 seconds.
 #> Optimizer convergence: successful.
 #> Running time for posterior draws 
-#> (1000 signal strength posterior draws per AE-drug pair):0.044 seconds.
+#> (1000 signal strength posterior draws per AE-drug pair):0.0424 seconds.
 #> 
 #> Extract estimated prior parameters, discovered signals
 #> and signal strength posterior draws using `summary()`.
 
 # fit K-gamma model with K = 3
 fit_Kgamma <- pvEBayes(
[…]
