+++ reference/summary.pvEBayes.md
@@ -74,20 +74,20 @@ obj <- pvEBayes(
   alpha = 0.5, n_posterior_draws = 10000
 )
 #> ℹ Fitting general-gamma model...
-#> ✔ Fitting general-gamma model... [314ms]
+#> ✔ Fitting general-gamma model... [310ms]
 #> 
 #> ℹ Generating 10000 posterior draws...
-#> ✔ Generating 10000 posterior draws... [406ms]
+#> ✔ Generating 10000 posterior draws... [400ms]
 #> 
 #> Object of class 'pvEBayes'
 #> 
 #> General-gamma model with hyperparameter alpha = 0.5.
 #> Estimated prior is a mixture of 18 gamma distributions.
 #> 
-#> Running time of the general-gamma model fitting: 0.3221 seconds.
+#> Running time of the general-gamma model fitting: 0.3174 seconds.
 #> Optimizer convergence: successful.
 #> Running time for posterior draws 
-#> (10000 signal strength posterior draws per AE-drug pair):0.4142 seconds.
+#> (10000 signal strength posterior draws per AE-drug pair):0.4081 seconds.
 #> 
 #> Extract estimated prior parameters, discovered signals
 #> and signal strength posterior draws using `summary()`.
@@ -99,37 +99,37 @@ summary(obj)
 #> General-gamma model with hyperparameter alpha = 0.5.
 #> Estimated prior is a mixture of 18 gamma distributions.
 #> 
-#> Running time of the general-gamma model fitting: 0.3221 seconds.
+#> Running time of the general-gamma model fitting: 0.3174 seconds.
 #> Optimizer convergence: successful.
 #> Running time for posterior draws 
-#> (10000 signal strength posterior draws per AE-drug pair):0.4142 seconds.
+#> (10000 signal strength posterior draws per AE-drug pair):0.4081 seconds.
 #> 
 #> Extract estimated prior parameters, discovered signals
 #> and signal strength posterior draws using `summary()`.
 #>                       AE         drug     N      E post_prob         q05
 #>                   <char>       <char> <int>  <num>     <num>       <num>
-#>   1: Acute Kidney Injury Atorvastatin  1132 532.74    1.0000  2.09795818
-#>   2: Acute Kidney Injury  Fluvastatin    23  50.91    0.0001  0.43557502
-#>   3: Acute Kidney Injury   Lovastatin    23   4.97    1.0000  2.76796628
-#>   4: Acute Kidney Injury  Pravastatin   153  74.39    1.0000  2.09621284
-#>   5: Acute Kidney Injury Rosuvastatin  1141 424.95    1.0000  2.72908781
+#>   1: Acute Kidney Injury Atorvastatin  1132 532.74    1.0000  2.09799790
+#>   2: Acute Kidney Injury  Fluvastatin    23  50.91    0.0001  0.43548758
+#>   3: Acute Kidney Injury   Lovastatin    23   4.97    1.0000  2.76671157
+#>   4: Acute Kidney Injury  Pravastatin   153  74.39    1.0000  2.09616011
+#>   5: Acute Kidney Injury Rosuvastatin  1141 424.95    1.0000  2.72871539
 #>  ---                                                                    
-#> 311:   Tendon Discomfort   Lovastatin     0   0.01    0.6029  0.10841846
-#> 312:   Tendon Discomfort  Pravastatin     0   0.08    0.5211  0.10599733
-#> 313:   Tendon Discomfort Rosuvastatin    10   0.45    1.0000 12.26673793
[…]
