### 2026-05-20 1bbbdfe
update CITATION
