+++ reference/logLik.pvEBayes.md
@@ -36,14 +36,14 @@ fit <- pvEBayes(
   alpha = 0.3, n_posterior_draws = NULL
 )
 #> ℹ Fitting general-gamma model...
-#> ✔ Fitting general-gamma model... [209ms]
+#> ✔ Fitting general-gamma model... [218ms]
 #> 
 #> Object of class 'pvEBayes'
 #> 
 #> General-gamma model with hyperparameter alpha = 0.3.
 #> Estimated prior is a mixture of 18 gamma distributions.
 #> 
-#> Running time of the general-gamma model fitting: 0.2167 seconds.
+#> Running time of the general-gamma model fitting: 0.2259 seconds.
 #> Optimizer convergence: successful.
 #> No posterior draws were generated.
 #> 
