+++ reference/plot.pvEBayes.md
@@ -41,20 +41,20 @@ A [ggplot](https://ggplot2.tidyverse.org/reference/ggplot.html) object.
 
 obj <- pvEBayes(statin2025_44, model = "general-gamma", alpha = 0.5)
 #> ℹ Fitting general-gamma model...
-#> ✔ Fitting general-gamma model... [310ms]
+#> ✔ Fitting general-gamma model... [325ms]
 #> 
 #> ℹ Generating 1000 posterior draws...
-#> ✔ Generating 1000 posterior draws... [34ms]
+#> ✔ Generating 1000 posterior draws... [36ms]
 #> 
 #> Object of class 'pvEBayes'
 #> 
 #> General-gamma model with hyperparameter alpha = 0.5.
 #> Estimated prior is a mixture of 18 gamma distributions.
 #> 
-#> Running time of the general-gamma model fitting: 0.3175 seconds.
+#> Running time of the general-gamma model fitting: 0.3323 seconds.
 #> Optimizer convergence: successful.
 #> Running time for posterior draws 
-#> (1000 signal strength posterior draws per AE-drug pair):0.0418 seconds.
+#> (1000 signal strength posterior draws per AE-drug pair):0.0435 seconds.
 #> 
 #> Extract estimated prior parameters, discovered signals
 #> and signal strength posterior draws using `summary()`.
